==> 01a_bag_of_words_cleaned.ipynb <==
import pandas as pd
import numpy as np
import nltk
import os
import re
import json
from bs4 import BeautifulSoup
from tqdm.auto import tqdm

#load dictionary: 
dict_df = pd.read_csv('data/loughran_mcdonald_word_list.csv')

dict_df.info()

import numpy as np
df = dict_df[(dict_df[['Negative','Positive', 'Uncertainty', 'Litigious', 'Strong_Modal', 'Weak_Modal', 'Constraining']] != 0).any(axis=1)]

def chang_to_ones(col): 
    if col != 0: 
        col = 1
    return col 

df['Negative']= df.Negative.apply(lambda x: chang_to_ones(x))
df['Positive']= df.Positive.apply(lambda x: chang_to_ones(x))
df['Uncertainty']= df.Uncertainty.apply(lambda x: chang_to_ones(x))
df['Litigious']= df.Litigious.apply(lambda x: chang_to_ones(x))
df['Strong_Modal']= df.Strong_Modal.apply(lambda x: chang_to_ones(x))
df['Weak_Modal']= df.Weak_Modal.apply(lambda x: chang_to_ones(x))
df['Constraining']= df.Constraining.apply(lambda x: chang_to_ones(x))


with open('test.txt', 'r') as file:
    tester = file.read().replace('\n', '')

# creates dictionary for each category: 
def create_dict(df, col): 
    d = list()
    # keep only relevant cols
    df= df[['Word', col]]
    for index, row in df.iterrows(): 
        if row[col]==1:
            d.append(row['Word'])   
        else:
            continue
    return d   

# 'Negative','Positive', 'Uncertainty', 'Litigious', 'Strong_Modal', 'Weak_Modal', 'Constraining']]

# create a list of words for each of the categories: 

# negative dictionary: 
negative_dict = create_dict(df, 'Negative') # 2355
# Positive dictionary: 
positive_dictionary = create_dict(df, 'Positive')#354
# len(positive_dictionary)
# uncertainity dictionary: 
uncertain_dict = create_dict(df, 'Uncertainty')# 297 
# len(uncertain_dict) 
#Litigious dictionary: 
litigious_dict = create_dict(df, 'Litigious')# 905 
# len(litigious_dict) 
# strong modal dicitonary: 
strong_modal_dict = create_dict(df, 'Strong_Modal')# 19 
# len(strong_modal_dict) 
# weak modal dicitonary: 
weak_modal_dict = create_dict(df, 'Weak_Modal')# 27
# len(weak_modal_dict)
# constraining dictionary 
constraining_dict = create_dict(df, 'Constraining')# 184
len(constraining_dict)

def get_weight(acc_nr, text): 
    text 
    negative_cnt = 0 
    positive_cnt = 0 
    uncertain_cnt = 0 
    litigious_cnt = 0
    strong_m_cnt = 0
    weak_m_cnt = 0 
    constraining_cnt = 0 
    
    # DICTIONARY of negative word counts: 
    negative_word_dict = dict()
    negative_word_prop= dict()
    
    positive_word_dict= dict()
    positive_word_prop= dict()
    
    uncertain_word_dict = dict()
    uncertain_word_prop = dict()
    litigious_word_dict = dict()
    litigious_word_prop= dict()
    strong_mod_word_dict = dict()
    strong_mod_word_prop= dict()
    weak_mod_word_dict= dict()
    weak_mod_word_prop= dict()
    constraining_word_dict= dict()
    constraining_word_prop= dict()    
    weight_dict = dict()
    
    doc_lst = nltk.wordpunct_tokenize(text.upper()) #list of all words in the document
    clean_list = list()
    
    # clean the document and remove unncecesary letters and numbers: 
    for item in doc_lst: 
        if item.isalpha() and ((len(item)>1) or item == 'A'): 
            clean_list.append(item)
    
    
    doc_len = len(clean_list) # length of the clean document 
    for item in clean_list:      
            if item in negative_dict: 
                negative_cnt+=1 # gives total negative cnt in the documents(not removing duplicates)
                #             if item not in negative.i: 
                if item not in negative_word_dict:
                    negative_word_dict[item] = clean_list.count(item)
                    negative_word_prop[item]= (negative_word_dict[item])/(doc_len)
                
            if item in positive_dictionary: 
                positive_cnt+=1
                if item not in positive_word_dict:
                    positive_word_dict[item] = clean_list.count(item)
                    positive_word_prop[item]= (positive_word_dict[item])/(doc_len)
                
            if item in uncertain_dict: 
                uncertain_cnt+=1
                if item not in uncertain_word_dict:
                    uncertain_word_dict[item] = clean_list.count(item)
                    uncertain_word_prop[item]= (uncertain_word_dict[item])/(doc_len)    
                
            if item in litigious_dict: 
                litigious_cnt+=1
                if item not in litigious_word_dict:
                    litigious_word_dict[item] = clean_list.count(item)
                    litigious_word_prop[item]= (litigious_word_dict[item])/(doc_len)
                
                
            if item in strong_modal_dict: 
                strong_m_cnt+=1
                if item not in strong_mod_word_dict:
                    strong_mod_word_dict[item] = clean_list.count(item)
                    strong_mod_word_prop[item]= (strong_mod_word_dict[item])/(doc_len)
                
            if item in weak_modal_dict: 
                weak_m_cnt+=1
                if item not in weak_mod_word_dict:
                    weak_mod_word_dict[item] = clean_list.count(item)
                    weak_mod_word_prop[item]= (weak_mod_word_dict[item])/(doc_len)
                
            if item in constraining_dict: 
                constraining_cnt+=1
                if item not in constraining_word_dict:
                    constraining_word_dict[item] = clean_list.count(item)
                    constraining_word_prop[item]= (constraining_word_dict[item])/(doc_len)
                
                
            else: 
                continue 
#       'acc_nr': acc_nr  
    weight_dict= {  'acc_nr': acc_nr, 'negative': negative_cnt, 'positive': positive_cnt, 'uncertain': uncertain_cnt,
                  'litigious':litigious_cnt, 'strong_m': strong_m_cnt, 'weak_m':weak_m_cnt,
                  'constraining': constraining_cnt, 'negative_freq':negative_word_dict, 'negative_prop':negative_word_prop,
                 'positive_freq':positive_word_dict, 'positive_prop': positive_word_prop, 
                  'uncertain_freq':uncertain_word_dict, 'uncertain_prop':uncertain_word_prop, 
                 'litigious_freq':litigious_word_dict, 'litigious_prop':litigious_word_prop, 
                 'strong_mod_freq':strong_mod_word_dict, 'strong_mod_prop':strong_mod_word_prop, 
                 'weak_mod_freq':weak_mod_word_dict, 'weak_mod_prop':weak_mod_word_prop, 
                 'constraining_freq':constraining_word_dict, 'constraining_prop':constraining_word_prop }
    
    return weight_dict
        

# Get list of old acc numbers so we don't have to run it through the json scrape again
acc_nr_store = json.load(open('metastore/master_output_2', 'r'))
old_acc_nr = []
for i in range(len(acc_nr_store)):
    old_acc_nr.append(acc_nr_store[i]['acc_nr'])
    

acc_nr_store = json.load(open("metastore/acc_nr_store.json", "r"))
def load_sec_filing(acc_nr, path):
    with open(os.path.join(path, f'{acc_nr}.html'), 'r') as f:
        contents = f.read()
    soup = BeautifulSoup(contents, 'lxml')#lxml
    return soup
    
def clean_filing(soup):
    # remove all table tags from document
    for s in soup.select("table"):
        s.extract()
        
    # only keep document tag (removes metadata)
    for s in soup.select("document"):
        document = s
        break
    document = document.get_text()
    document = re.sub(r"[\n|\s]", " ", document)
    document = re.sub(r"[^\s|\w|\d]", " ", document)
    return document


bag_of_words_store = list()
for cik in acc_nr_store:
    # sentiment_store[cik] = {}
    for acc_nr in tqdm(acc_nr_store[cik]):
            if acc_nr not in old_acc_nr: 
                soup = load_sec_filing(acc_nr, "metastore/filings")
                document = clean_filing(soup)
                bag_of_words_store.append(get_weight(acc_nr, document))
            else: 
                continue    

with open('metastore/master_output_3', 'w') as fout:
    json.dump(bag_of_words_store, fout)

# 2015-2021 data
import json

with open('metastore/master_output_2') as f:
    raw_master_output = json.load(f)

# 2001-2015 data
import json

with open('metastore/master_output_3') as f:
    raw_master_output = json.load(f)

### Calculate word count of 10-K, total proportion of negative words, expected average negative ###
### This calculation is performed for each 10-K filing
for i in range(len(raw_master_output)):
    
    if len(list(raw_master_output[i]['negative_freq'].keys())) != 0:
        first_key_negative_freq = list(raw_master_output[i]['negative_freq'].keys())[0]
        first_key_negative_prop = list(raw_master_output[i]['negative_prop'].keys())[0]
        
        doc_len_individual_10k = raw_master_output[i]['negative_freq'][first_key_negative_freq] / raw_master_output[i]['negative_prop'][first_key_negative_prop]
        raw_master_output[i]['doc_len'] = round(doc_len_individual_10k)
        
    elif len(list(raw_master_output[i]['positive_freq'].keys())) != 0:
        first_key_positive_freq = list(raw_master_output[i]['positive_freq'].keys())[0]
        first_key_positive_prop = list(raw_master_output[i]['positive_prop'].keys())[0]
        
        doc_len_individual_10k = raw_master_output[i]['positive_freq'][ first_key_positive_freq] / raw_master_output[i]['positive_prop'][first_key_positive_prop]
        raw_master_output[i]['doc_len'] = round(doc_len_individual_10k)
        
    elif len(list(raw_master_output[i]['uncertain_freq'].keys())) != 0:
        first_key_uncertain_freq = list(raw_master_output[i]['uncertain_freq'].keys())[0]
        first_key_uncertain_prop = list(raw_master_output[i]['uncertain_prop'].keys())[0]
        
        doc_len_individual_10k = raw_master_output[i]['uncertain_freq'][first_key_uncertain_fre] / raw_master_output[i]['uncertain_prop'][first_key_uncertain_prop]
        raw_master_output[i]['doc_len'] = round(doc_len_individual_10k)
    
    elif len(list(raw_master_output[i]['litigious_freq'].keys())) != 0:
        first_key_litigious_freq = list(raw_master_output[i]['litigious_freq'].keys())[0]
        first_key_litigious_prop = list(raw_master_output[i]['litigious_prop'].keys())[0]
        
        doc_len_individual_10k = raw_master_output[i]['litigious_freq'][first_key_litigious_freq] / raw_master_output[i]['litigious_prop'][first_key_litigious_prop]
        raw_master_output[i]['doc_len'] = round(doc_len_individual_10k)
        
    elif len(list(raw_master_output[i]['strong_mod_freq'].keys())) != 0:
        first_key_strong_m_freq = list(raw_master_output[i]['strong_mod_freq'].keys())[0]
        first_key_strong_m_prop = list(raw_master_output[i]['strong_mod_prop'].keys())[0]
        
        doc_len_individual_10k = raw_master_output[i]['strong_mod_freq'][first_key_strong_m_freq] / raw_master_output[i]['strong_mod_prop'][first_key_strong_m_prop]
        raw_master_output[i]['doc_len'] = round(doc_len_individual_10k)
        
    elif len(list(raw_master_output[i]['weak_mod_freq'].keys())) != 0:
        first_key_weak_m_freq = list(raw_master_output[i]['weak_mod_freq'].keys())[0]
        first_key_weak_m_prop = list(raw_master_output[i]['weak_mod_prop'].keys())[0]
        
        doc_len_individual_10k = raw_master_output[i]['weak_mod_freq'][first_key_weak_m_freq] / raw_master_output[i]['weak_mod_prop'][first_key_weak_m_prop]
        raw_master_output[i]['doc_len'] = round(doc_len_individual_10k)
        
    elif len(list(raw_master_output[i]['constraining_freq'].keys())) != 0:
        first_key_constraining_freq = list(raw_master_output[i]['constraining_freq'].keys())[0]
        first_key_constraining_prop = list(raw_master_output[i]['constraining_prop'].keys())[0]
        
        doc_len_individual_10k = raw_master_output[i]['constraining_freq'][first_key_constraining_freq] / raw_master_output[i]['constraining_prop'][first_key_constraining_prop]
        raw_master_output[i]['doc_len'] = round(doc_len_individual_10k)
    
    else:
        raw_master_output[i]['doc_len'] = 0
    
    # Calculation of total proportion of each word label
    # Negative
    if raw_master_output[i]['negative'] == 0:
        raw_master_output[i]['total_negative_proportion'] = 0
    if raw_master_output[i]['negative'] != 0:
        raw_master_output[i]['total_negative_proportion'] = raw_master_output[i]['negative'] / raw_master_output[i]['doc_len']
        
    # Positive
    if raw_master_output[i]['positive'] == 0:
        raw_master_output[i]['total_positive_proportion'] = 0
    if  raw_master_output[i]['positive'] != 0:
        raw_master_output[i]['total_positive_proportion'] = raw_master_output[i]['positive'] / raw_master_output[i]['doc_len']
        
    # Uncertain       
    if raw_master_output[i]['uncertain'] == 0:
        raw_master_output[i]['total_uncertain_proportion'] = 0
    if raw_master_output[i]['uncertain'] != 0:
        raw_master_output[i]['total_uncertain_proportion'] = raw_master_output[i]['uncertain'] / raw_master_output[i]['doc_len']
    
    # Litigious
    if raw_master_output[i]['litigious'] == 0:
        raw_master_output[i]['total_litigious_proportion'] = 0
    if raw_master_output[i]['litigious'] != 0:
        raw_master_output[i]['total_litigious_proportion'] = raw_master_output[i]['litigious'] / raw_master_output[i]['doc_len']
    
    # Strong_m
    if raw_master_output[i]['strong_m'] == 0:
        raw_master_output[i]['total_strong_m_proportion'] = 0
    if raw_master_output[i]['strong_m'] != 0:
        raw_master_output[i]['total_strong_m_proportion'] = raw_master_output[i]['strong_m'] / raw_master_output[i]['doc_len']
    
    # Weak_m
    if raw_master_output[i]['weak_m'] == 0:
        raw_master_output[i]['total_weak_m_proportion'] = 0
    if raw_master_output[i]['weak_m'] != 0:
        raw_master_output[i]['total_weak_m_proportion'] = raw_master_output[i]['weak_m'] / raw_master_output[i]['doc_len']
    
    # Constraining
    if raw_master_output[i]['constraining'] == 0:
         raw_master_output[i]['total_constraining_proportion'] = 0
    if raw_master_output[i]['constraining'] != 0:
        raw_master_output[i]['total_constraining_proportion'] = raw_master_output[i]['constraining'] / raw_master_output[i]['doc_len']
    
    # Calculation of average value for each word label
    # Negative
    if len(list(raw_master_output[i]['negative_freq'].keys())) != 0:
        expected_average_negative = raw_master_output[i]['negative'] / len(list(raw_master_output[i]['negative_freq'].keys()))
        raw_master_output[i]['expected_average_negative'] = expected_average_negative
    else:
        raw_master_output[i]['expected_average_negative'] = 0
    
    # Positive
    if len(list(raw_master_output[i]['positive_freq'].keys())) != 0:
        expected_average_positive = raw_master_output[i]['positive'] / len(list(raw_master_output[i]['positive_freq'].keys()))
        raw_master_output[i]['expected_average_positive'] = expected_average_positive
    else: 
        raw_master_output[i]['expected_average_positive'] = 0
        
    # Uncertain
    if len(list(raw_master_output[i]['uncertain_freq'].keys())) != 0:
        expected_average_uncertain = raw_master_output[i]['uncertain'] / len(list(raw_master_output[i]['uncertain_freq'].keys()))
        raw_master_output[i]['expected_average_uncertain'] = expected_average_uncertain
    else:
        raw_master_output[i]['expected_average_uncertain'] = 0
    
    # Litigious
    if len(list(raw_master_output[i]['litigious_freq'].keys())) != 0:
        expected_average_litigious = raw_master_output[i]['litigious'] / len(list(raw_master_output[i]['litigious_freq'].keys()))
        raw_master_output[i]['expected_average_litigious'] = expected_average_litigious
    else: 
        raw_master_output[i]['expected_average_litigious'] = 0
        
    # Strong_m
    if len(list(raw_master_output[i]['strong_mod_freq'].keys())) != 0:
        expected_average_strong_m = raw_master_output[i]['strong_m'] / len(list(raw_master_output[i]['strong_mod_freq'].keys()))
        raw_master_output[i]['expected_average_strong_m'] = expected_average_strong_m
    else:
        raw_master_output[i]['expected_average_strong_m'] = 0
    
    # Weak_m
    if len(list(raw_master_output[i]['weak_mod_freq'].keys())) != 0:
        expected_average_weak_m = raw_master_output[i]['weak_m'] / len(list(raw_master_output[i]['weak_mod_freq'].keys()))
        raw_master_output[i]['expected_average_weak_m'] = expected_average_weak_m
    else: 
        raw_master_output[i]['expected_average_weak_m'] = 0
    
    # Constraining
    if len(list(raw_master_output[i]['constraining_freq'].keys())) != 0:
        expected_average_constraining = raw_master_output[i]['constraining'] / len(list(raw_master_output[i]['constraining_freq'].keys()))
        raw_master_output[i]['expected_average_constraining'] = expected_average_constraining
    else:
        raw_master_output[i]['expected_average_constraining'] = 0


# Original 2001-2015 data had 1318 10-K's, after filtering we came to 1291 10-K's
master_output = []
for i in range(len(raw_master_output)):
    if raw_master_output[i]['doc_len'] > 2000:
        master_output.append(raw_master_output[i])


### Calculate number of occurrence of each word in all the 10-K filings ###
negative_word_list = list()
positive_word_list = list()
uncertain_word_list = list()
litigious_word_list = list()
strong_m_word_list = list()
weak_m_word_list= list()
constraining_word_list= list()


for i in range(len(master_output)): 
    temp = list(master_output[i]['negative_freq'].keys())
    negative_word_list.extend(temp) # contains all negative words in all 10k filings
    
    positive_temp = list(master_output[i]['positive_freq'].keys())
    positive_word_list.extend(positive_temp)
    
    # uncertain: 
    uncertain_temp = list(master_output[i]['uncertain_freq'].keys())
    uncertain_word_list.extend(uncertain_temp)
    
    #litigious: 
    litigious_temp = list(master_output[i]['litigious_freq'].keys())
    litigious_word_list.extend(litigious_temp)
    
    #strong m: 
    strong_m_temp = list(master_output[i]['strong_mod_freq'].keys())
    strong_m_word_list.extend(strong_m_temp)
    
    #weak m:
    weak_m_temp = list(master_output[i]['weak_mod_freq'].keys())
    weak_m_word_list.extend(weak_m_temp)
    
    # constraining: 
    constraining_temp = list(master_output[i]['constraining_freq'].keys())
    constraining_word_list.extend(constraining_temp)
        

# Create dictionary of counts of each negative word across 10-K filings
df_dict = dict()
for item in negative_word_list: 
    if item not in df_dict: 
        df_dict[item]= negative_word_list.count(item)
        
        
        
#POSITIVE: Create dictionary of counts of each negative word across 10-K filings
positive_df_dict = dict()
for item in positive_word_list: 
    if item not in positive_df_dict: 
        positive_df_dict[item]= positive_word_list.count(item)
        
#uncertain: Create dictionary of counts of each uncertain word across 10-K filings
uncertain_df_dict = dict()
for item in uncertain_word_list: 
    if item not in uncertain_df_dict: 
        uncertain_df_dict[item]= uncertain_word_list.count(item)  

#litigious: Create dictionary of counts of each litigious word across 10-K filings
litigious_df_dict = dict()
for item in litigious_word_list: 
    if item not in litigious_df_dict: 
        litigious_df_dict[item]= litigious_word_list.count(item)
        
#strongm: Create dictionary of counts of each strongm word across 10-K filings
strong_m_df_dict = dict()
for item in strong_m_word_list: 
    if item not in strong_m_df_dict: 
        strong_m_df_dict[item]= strong_m_word_list.count(item)
        

#weakm: Create dictionary of counts of each weakm word across 10-K filings
weak_m_df_dict = dict()
for item in weak_m_word_list: 
    if item not in weak_m_df_dict: 
        weak_m_df_dict[item]= weak_m_word_list.count(item)     
        

#constraining: Create dictionary of counts of each weakm word across 10-K filings
constraining_df_dict = dict()
for item in constraining_word_list: 
    if item not in constraining_df_dict: 
        constraining_df_dict[item]= constraining_word_list.count(item) 
        
###########################################################################################################################
#1 Added df_dict to each 10-K filing (only added words found in 10-K)
for i in range(len(master_output)):
    master_output[i]['negative_df_dict'] = dict()
    temp = list(master_output[i]['negative_freq'].keys())
    temp_df_dict = list(df_dict.keys())
    for key in temp_df_dict:
        if key in temp:
            master_output[i]['negative_df_dict'][key] = df_dict[key]
#2. 
for i in range(len(master_output)):
    master_output[i]['positive_df_dict'] = dict()
    temp = list(master_output[i]['positive_freq'].keys())
    temp_df_dict = list(positive_df_dict.keys())
    for key in temp_df_dict:
        if key in temp:
            master_output[i]['positive_df_dict'][key] = positive_df_dict[key]

            
#3.uncertain_df_dict
for i in range(len(master_output)):
    master_output[i]['uncertain_df_dict'] = dict()
    temp = list(master_output[i]['uncertain_freq'].keys())
    temp_df_dict = list(uncertain_df_dict.keys())
    for key in temp_df_dict:
        if key in temp:
            master_output[i]['uncertain_df_dict'][key] = uncertain_df_dict[key]

#4.litigious_df_dict
for i in range(len(master_output)):
    master_output[i]['litigious_df_dict'] = dict()
    temp = list(master_output[i]['litigious_freq'].keys())
    temp_df_dict = list(litigious_df_dict.keys())
    for key in temp_df_dict:
        if key in temp:
            master_output[i]['litigious_df_dict'][key] = litigious_df_dict[key]

#5.strong_mod_freq
for i in range(len(master_output)):
    master_output[i]['strong_m_df_dict'] = dict()
    temp = list(master_output[i]['strong_mod_freq'].keys())
    temp_df_dict = list(strong_m_df_dict.keys())
    for key in temp_df_dict:
        if key in temp:
            master_output[i]['strong_m_df_dict'][key] = strong_m_df_dict[key]
#6.weak_mod_freq
for i in range(len(master_output)):
    master_output[i]['weak_m_df_dict'] = dict()
    temp = list(master_output[i]['weak_mod_freq'].keys())
    temp_df_dict = list(weak_m_df_dict.keys())
    for key in temp_df_dict:
        if key in temp:
            master_output[i]['weak_m_df_dict'][key] = weak_m_df_dict[key]
            
#7.constraining_df_dict constraining_freq
for i in range(len(master_output)):
    master_output[i]['constraining_df_dict'] = dict()
    temp = list(master_output[i]['constraining_freq'].keys())
    temp_df_dict = list(constraining_df_dict.keys())
    for key in temp_df_dict:
        if key in temp:
            master_output[i]['constraining_df_dict'][key] = constraining_df_dict[key]
###########################################################################################################################            
# # Reorder negative_freq & df_dict alphabetically in each 10-K filing
for i in range(len(master_output)):
    master_output[i]['negative_freq'] = dict(sorted(master_output[0]['negative_freq'].items(), key = lambda x: x[0], reverse = False))
    master_output[i]['negative_df_dict'] = dict(sorted(master_output[0]['negative_df_dict'].items(), key = lambda x: x[0], reverse = False))
    
    #positive: 
    master_output[i]['positive_freq'] = dict(sorted(master_output[0]['positive_freq'].items(), key = lambda x: x[0], reverse = False))
    master_output[i]['positive_df_dict'] = dict(sorted(master_output[0]['positive_df_dict'].items(), key = lambda x: x[0], reverse = False))
    
    # uncertain: 
    master_output[i]['uncertain_freq'] = dict(sorted(master_output[0]['uncertain_freq'].items(), key = lambda x: x[0], reverse = False))
    master_output[i]['uncertain_df_dict'] = dict(sorted(master_output[0]['uncertain_df_dict'].items(), key = lambda x: x[0], reverse = False))
    
    #litigious: 
    master_output[i]['litigious_freq'] = dict(sorted(master_output[0]['litigious_freq'].items(), key = lambda x: x[0], reverse = False))
    master_output[i]['litigious_df_dict'] = dict(sorted(master_output[0]['litigious_df_dict'].items(), key = lambda x: x[0], reverse = False))
    
    #strong_mod_freq
    
    master_output[i]['strong_mod_freq'] = dict(sorted(master_output[0]['strong_mod_freq'].items(), key = lambda x: x[0], reverse = False))
    master_output[i]['strong_m_df_dict'] = dict(sorted(master_output[0]['strong_m_df_dict'].items(), key = lambda x: x[0], reverse = False))
    
    # weak mod:
    master_output[i]['weak_mod_freq'] = dict(sorted(master_output[0]['weak_mod_freq'].items(), key = lambda x: x[0], reverse = False))
    master_output[i]['weak_m_df_dict'] = dict(sorted(master_output[0]['weak_m_df_dict'].items(), key = lambda x: x[0], reverse = False))
    
    
    # constraining
    master_output[i]['constraining_freq'] = dict(sorted(master_output[0]['constraining_freq'].items(), key = lambda x: x[0], reverse = False))
    master_output[i]['constraining_df_dict'] = dict(sorted(master_output[0]['constraining_df_dict'].items(), key = lambda x: x[0], reverse = False))

<div class="alert alert-block alert-info">
w_i_j =[ ( 1 + log(tf_i_j) ) / ( 1 + log(a_j) ) ] * log(N/df_i)
</div>

### Calculate weight of each word in each 10-K filing ###

N = len(master_output)

# for i in range(len(master_output)):
#     negative_word_list_per_10k = list(master_output[i]['negative_freq'].keys())
#     master_output[i]['weight_dict'] = dict()
    
#     for word in negative_word_list_per_10k:
#         numerator = 1 + np.log(master_output[i]['negative_freq'][word])
#         denominator = 1 + np.log(master_output[i]['expected_average_negative'])
#         multiplication_term = np.log(N/master_output[i]['df_dict'][word]) # Multiply by log(N/number of times that word shows up across 10-K)
        
#         master_output[i]['weight_dict'][word] = (numerator / denominator) * multiplication_term
#         master_output[i]['summed_negative_weights'] = sum(list(master_output[i]['weight_dict'].values())) # Final weight calculation

for i in range(len(master_output)):
    
    negative_word_list_per_10k = list(master_output[i]['negative_freq'].keys())
    master_output[i]['weight_dict_negative'] = dict()
    
    positive_word_list_per_10k = list(master_output[i]['positive_freq'].keys())
    master_output[i]['weight_dict_positive'] = dict()
    
    uncertain_word_list_per_10k = list(master_output[i]['uncertain_freq'].keys())
    master_output[i]['weight_dict_uncertain'] = dict()
    
    litigious_word_list_per_10k = list(master_output[i]['litigious_freq'].keys())
    master_output[i]['weight_dict_litigious'] = dict()
    
    strong_m_word_list_per_10k = list(master_output[i]['strong_mod_freq'].keys())
    master_output[i]['weight_dict_strong_m'] = dict()
    
    weak_m_word_list_per_10k = list(master_output[i]['weak_mod_freq'].keys())
    master_output[i]['weight_dict_weak_m'] = dict()
    
    constraining_word_list_per_10k = list(master_output[i]['constraining_freq'].keys())
    master_output[i]['weight_dict_constraining'] = dict()
    
    # Negative
    for word in negative_word_list_per_10k:
        numerator = 1 + np.log(master_output[i]['negative_freq'][word])
        denominator = 1 + np.log(master_output[i]['expected_average_negative'])
        multiplication_term = np.log(N/master_output[i]['negative_df_dict'][word]) # Multiply by log(N/number of times that word shows up across 10-K)
        
        master_output[i]['weight_dict_negative'][word] = (numerator / denominator) * multiplication_term
        master_output[i]['summed_negative_weights'] = sum(list(master_output[i]['weight_dict_negative'].values())) # Final weight calculation
    
    # Positive
    for word in positive_word_list_per_10k:
        numerator = 1 + np.log(master_output[i]['positive_freq'][word])
        denominator = 1 + np.log(master_output[i]['expected_average_positive'])
        multiplication_term = np.log(N/master_output[i]['positive_df_dict'][word]) # Multiply by log(N/number of times that word shows up across 10-K)
        
        master_output[i]['weight_dict_positive'][word] = (numerator / denominator) * multiplication_term
        master_output[i]['summed_positive_weights'] = sum(list(master_output[i]['weight_dict_positive'].values())) # Final weight calculation

    # Uncertain
    for word in uncertain_word_list_per_10k:
        numerator = 1 + np.log(master_output[i]['uncertain_freq'][word])
        denominator = 1 + np.log(master_output[i]['expected_average_uncertain'])
        multiplication_term = np.log(N/master_output[i]['uncertain_df_dict'][word]) # Multiply by log(N/number of times that word shows up across 10-K)
        
        master_output[i]['weight_dict_uncertain'][word] = (numerator / denominator) * multiplication_term
        master_output[i]['summed_uncertain_weights'] = sum(list(master_output[i]['weight_dict_uncertain'].values())) # Final weight calculation

    # Litigious
    for word in litigious_word_list_per_10k:
        numerator = 1 + np.log(master_output[i]['litigious_freq'][word])
        denominator = 1 + np.log(master_output[i]['expected_average_litigious'])
        multiplication_term = np.log(N/master_output[i]['litigious_df_dict'][word]) # Multiply by log(N/number of times that word shows up across 10-K)
        
        master_output[i]['weight_dict_litigious'][word] = (numerator / denominator) * multiplication_term
        master_output[i]['summed_litigious_weights'] = sum(list(master_output[i]['weight_dict_litigious'].values())) # Final weight calculation
   
    # Strong_m
    for word in strong_m_word_list_per_10k:
        numerator = 1 + np.log(master_output[i]['strong_mod_freq'][word])
        denominator = 1 + np.log(master_output[i]['expected_average_strong_m'])
        multiplication_term = np.log(N/master_output[i]['strong_m_df_dict'][word]) # Multiply by log(N/number of times that word shows up across 10-K)
        
        master_output[i]['weight_dict_strong_m'][word] = (numerator / denominator) * multiplication_term
        master_output[i]['summed_strong_m_weights'] = sum(list(master_output[i]['weight_dict_strong_m'].values())) # Final weight calculation

    # Weak_m
    for word in weak_m_word_list_per_10k:
        numerator = 1 + np.log(master_output[i]['weak_mod_freq'][word])
        denominator = 1 + np.log(master_output[i]['expected_average_weak_m'])
        multiplication_term = np.log(N/master_output[i]['weak_m_df_dict'][word]) # Multiply by log(N/number of times that word shows up across 10-K)
        
        master_output[i]['weight_dict_weak_m'][word] = (numerator / denominator) * multiplication_term
        master_output[i]['summed_weak_m_weights'] = sum(list(master_output[i]['weight_dict_weak_m'].values())) # Final weight calculation

    # Constraining
    for word in constraining_word_list_per_10k:
        numerator = 1 + np.log(master_output[i]['constraining_freq'][word])
        denominator = 1 + np.log(master_output[i]['expected_average_negative'])
        multiplication_term = np.log(N/master_output[i]['constraining_df_dict'][word]) # Multiply by log(N/number of times that word shows up across 10-K)
        
        master_output[i]['weight_dict_constraining'][word] = (numerator / denominator) * multiplication_term
        master_output[i]['summed_constraining_weights'] = sum(list(master_output[i]['weight_dict_constraining'].values())) # Final weight calculation


# Final dictionary

final_list = list()
for i in range(len(master_output)):
    temp = dict()
    temp['acc_nr'] = master_output[i]['acc_nr']
    temp['total_negative_proportion'] = master_output[i]['total_negative_proportion']
    temp['total_positive_proportion'] = master_output[i]['total_positive_proportion']
    temp['total_uncertain_proportion'] = master_output[i]['total_uncertain_proportion']
    temp['total_litigious_proportion'] = master_output[i]['total_litigious_proportion']
    temp['total_strong_m_proportion'] = master_output[i]['total_strong_m_proportion']
    temp['total_weak_m_proportion'] = master_output[i]['total_weak_m_proportion']
    temp['total_constraining_proportion'] = master_output[i]['total_constraining_proportion']
    
    temp['summed_negative_weights'] = master_output[i]['summed_negative_weights']
    temp['summed_positive_weights'] = master_output[i]['summed_positive_weights']
    temp['summed_uncertain_weights'] = master_output[i]['summed_uncertain_weights']
    temp['summed_litigious_weights'] = master_output[i]['summed_litigious_weights']
    temp['summed_strong_m_weights'] = master_output[i]['summed_strong_m_weights']
    temp['summed_weak_m_weights'] = master_output[i]['summed_weak_m_weights']
    temp['summed_constraining_weights'] = master_output[i]['summed_constraining_weights']
    final_list.append(temp)
    
# Final dataframe
final_df = pd.DataFrame(final_list)

final_df.to_csv('metastore/bag_of_words_2001-2015.csv', index = False)

# Union of 2001-2015 & 2015-2021 Data

old_2001_2015 = pd.read_csv('metastore/bag_of_words_2001-2015.csv')
new_2015_2021 = pd.read_csv('metastore/bag_of_words_2015-2021.csv')

combined_data = pd.concat([old_2001_2015, new_2015_2021])
combined_data.to_csv('metastore/bag_of_words_2001-2021.csv', index = False)
